+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * MPNumericAux directionOffsetToDegrees() and normalFromTangent() returning wrong values for some directions.
   * Fixing popcorn point type translation. [reported by Daniel Bean]
   * LSize and Orientation arrows not presented when end point is selected.
+  * Interactive Orientation arrow not moving with azimuth picker arrow is moved.
 * Infrastructure maintenance:
   * Testing all possible variations of scrap scale option parsing.
   * Custom exceptions should provide trace.